DailyMe/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
pendulum==2.1.2